Update Fetch Product List API

+++ app/src/main/java/com/example/frontend/ProductData.kt
@@ -1,14 +1,16 @@
 package com.example.frontend
 
 import java.io.Serializable
+import java.util.Date
 
 data class ProductData  (
     val id : Int,
     val name : String,
     val inform : String,
     val img : String,
+    // val date: Date,
     val price : Int,
-    val xSize : Int,
-    val ySize : Int,
-    val zSize : Int
+    val xSize : Double,
+    val ySize : Double,
+    val zSize : Double
 ) : Serializable
